improvement(utils): ensure no longer takes an EnsureCreateOptions object as a second argument
BREAKING CHANGE: `ensure` used to take an EnsureCreateOptions object as a second argument allowing
to specify the input types that, when not an error, would be casted into a string for the new Error
to have as message. The new implementation casts all non Error input types into a string to use as a
message. Otherwise, passing an Error creating function as a second argument remains possible.